refactor: refact code modularization
refacor : 코드 기능별 모듈화
